requirements.txt
chromadb
fastapi 
llama-index
llama-index-vector-stores-chroma 
llama-index-embeddings-openai
pydantic
python-dotenv
uvicorn 
streamlit
python-multipart
loguru